refactor(ir): use pointer-based variable identity in SSA conversion
Replace name_hint_-keyed maps with const Var* pointer-keyed maps in the
ConvertToSSA pass, fixing a layering violation where a cosmetic string
was used as variable identity. Update the parser to reuse Var objects on
reassignment, and add clone_def_vars flag to DeepClone so UnrollLoops
preserves shared Var pointers across iterations.

+++ tests/ut/codegen/test_pto_codegen_paged_attn.py
@@ -120,7 +120,7 @@ class PagedAttention:
         pij_tile = pl.tile.exp(pij_tile)
         pij_bf16_tile = pl.tile.cast(pij_tile, mode="round", target_type=pl.BF16)
         pij_tile = pl.tile.cast(pij_bf16_tile, mode="round", target_type=pl.FP16)
-        sum_tile: pl.Tile[[16, 1], pl.FP16] = pl.tile.row_sum(pij_tile, tmp_tile)
+        sum_tile: pl.Tile[[16, 1], pl.BF16] = pl.tile.row_sum(pij_tile, tmp_tile)
         pl.store(max_tile, [0, 0], mij)
         pl.store(sum_tile, [0, 0], lij)
         pl.store(pij_bf16_tile, [0, 0], pij)
